Move some PCollectionView bits out of util
